Overwrite layout HTML and add favicons
A new layout HTML file was added to '_layout/default.html' to overwrite
the default theme HTML. The default.html was altered to include favicon
metadata and change built-in wording.

Additional github options were added to the _config.yml to add more
layout features.

Front-matter was added to contributors section.